+++ lab1/lab1/alphabet.h
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+void translate(int numIntervals[], char * alphabet[], char * lingChain[]);
+
+void printLingChain(char * lingChain[]);
+
