+++ lab1/lab1/translator.h
@@ -1,4 +1,2 @@
 #include <stdio.h>
 
-void translate(int numIntervals[], char * alphabet[], char * lingChain[]);
-void printLingChain(char * lingChain[]);
