+++ lab1/lab1/utils.h
@@ -1,0 +1,5 @@
+#include <stdio.h>
+
+int intArrLength(int arr[]);
+
+int charArrLength(char * arr[]);
